code cleanup for -Wconversion

+++ wcd/src/c3po/WcdStack.c
@@ -337,15 +337,15 @@ expfun void printWcdStack(text Offset,
       {
          if ((isEmptyWcdStackDir(w) == false) || (showEmpty == true))
          {
-            fprintf(fp, "%sint size : %zu\n", new_Offset, w->size);
+            fprintf(fp, "%sint size : %lu\n", new_Offset, (unsigned long)w->size);
 
             index = 0;
             while(index < w->size)
             {
                if (w->dir[index] ne NULL)
-               fprintf(fp, "%stext dir[%zu] : %s\n", new_Offset, index, w->dir[index]);
+               fprintf(fp, "%stext dir[%lu] : %s\n", new_Offset, (unsigned long)index, w->dir[index]);
                else if (showEmpty == true)
-                  fprintf(fp, "%stext dir[%zu] : NULL\n", new_Offset, index);
+                  fprintf(fp, "%stext dir[%lu] : NULL\n", new_Offset, (unsigned long)index);
                index = index + 1;
             }
          }
